requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.5
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-CORS==4.0.0

Werkzeug==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0

SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

bcrypt==4.1.2
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0

opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.26.2
face-recognition==1.3.5
dlib==19.24.2

scikit-learn==1.3.2
scipy==1.11.4

redis==5.0.1
celery==5.3.4
APScheduler==3.10.4

gunicorn==21.2.0
python-socketio==5.10.0
eventlet==0.33.3

prometheus-client==0.19.0
flask-swagger-ui==4.11.1

pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
